backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0

pandas==2.3.3
numpy==2.4.0
scikit-learn==1.5.0
scipy==1.16.3
joblib==1.4.2

firebase-admin==6.5.0
python-dotenv==1.2.1

requests==2.32.5
python-multipart==0.0.21

twilio==9.0.4
